+++ libs/designer/src/lib/ui/panel/panelTabs/parametersTab/identityselector.tsx
@@ -7,10 +7,9 @@ import type { IDropdownOption, IDropdownStyles } from '@fluentui/react';
 import { Dropdown, FontSizes, Label } from '@fluentui/react';
 import { WorkflowService } from '@microsoft/designer-client-services-logic-apps';
 import type { ManagedIdentity } from '@microsoft/utils-logic-apps';
-import { isIdentityAssociatedWithLogicApp, ResourceIdentityType, equals } from '@microsoft/utils-logic-apps';
+import { isIdentityAssociatedWithLogicApp, equals, getIdentityDropdownOptions } from '@microsoft/utils-logic-apps';
 import type { FormEvent } from 'react';
 import { useEffect, useState } from 'react';
-import type { IntlShape } from 'react-intl';
 import { useIntl } from 'react-intl';
 import { useDispatch, useSelector } from 'react-redux';
 
@@ -43,7 +42,7 @@ export const IdentitySelector = (props: IdentitySelectorProps) => {
   const intl = useIntl();
   const dispatch = useDispatch<AppDispatch>();
   const identity = WorkflowService().getAppIdentity?.() as ManagedIdentity;
-  const options = getDropdownOptions(identity, intl);
+  const options = getIdentityDropdownOptions(identity, intl);
 
   const selectedIdentity = useSelector((state: RootState) => {
     const { connectionProperties } = getConnectionReference(state.connections, nodeId);
@@ -108,25 +107,4 @@ export const IdentitySelector = (props: IdentitySelectorProps) => {
       />
     </div>
   );
-};
-
-const getDropdownOptions = (managedIdentity: ManagedIdentity, intl: IntlShape): IDropdownOption[] => {
-  const options: IDropdownOption[] = [];
-  const { type, userAssignedIdentities } = managedIdentity;
-  const systemAssigned = intl.formatMessage({
-    defaultMessage: 'System-assigned managed identity',
-    description: 'Text for system assigned managed identity',
-  });
-
-  if (equals(type, ResourceIdentityType.SYSTEM_ASSIGNED) || equals(type, ResourceIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED)) {
-    options.push({ key: constants.SYSTEM_ASSIGNED_MANAGED_IDENTITY, text: systemAssigned });
-  }
-
-  if (equals(type, ResourceIdentityType.USER_ASSIGNED) || equals(type, ResourceIdentityType.SYSTEM_ASSIGNED_USER_ASSIGNED)) {
-    for (const identity of Object.keys(userAssignedIdentities ?? {})) {
-      options.push({ key: identity, text: identity.split('/').at(-1) ?? identity });
-    }
-  }
-
-  return options;
 };
